Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.1.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.6.0
groq==0.23.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-groq==0.3.2
langchain-openai==0.1.8
langchain-text-splitters==0.3.8
langsmith==0.3.39
markdown-it-py==3.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
motor==3.3.2
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.2
passlib==1.7.4
propcache==0.3.1
PyMuPDF==1.25.1
pymongo==4.8.1
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.4
setuptools==78.1.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.7.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
yarl==1.20.0
zstandard==0.23.0
